+++ src/Transformer/ObjectToArrayTransformer.php
@@ -14,7 +14,9 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper\Transformer;
 
 use Rekalogika\Mapper\Context\Context;
+use Rekalogika\Mapper\Context\ContextMemberNotFoundException;
 use Rekalogika\Mapper\Exception\InvalidArgumentException;
+use Rekalogika\Mapper\Serializer\NormalizerContext;
 use Rekalogika\Mapper\Transformer\Contracts\TransformerInterface;
 use Rekalogika\Mapper\Transformer\Contracts\TypeMapping;
 use Rekalogika\Mapper\Util\TypeFactory;
@@ -44,9 +46,16 @@ final class ObjectToArrayTransformer implements TransformerInterface
             throw new InvalidArgumentException(sprintf('Source must be object, "%s" given', get_debug_type($source)), context: $context);
         }
 
+        try {
+            $normalizerContext = $context(NormalizerContext::class)->toArray();
+        } catch (ContextMemberNotFoundException) {
+            $normalizerContext = [];
+        }
+
         return $this->normalizer->normalize(
             $source,
             $this->normalizerFormat,
+            $normalizerContext
         );
     }